Cleared the domain files.
