+++ consensus/proposed_block.go
@@ -9,6 +9,7 @@ var ErrDecodingEmptyBlock = errors.New("Empty Block decoding failed")
 
 type ProposedBlock struct {
 	Seal []byte
+	body []byte
 }
 
 func (block *ProposedBlock) Serialize() ([]byte, error) {
